+++ NEWS.md
@@ -1,6 +1,7 @@
 # spant 2.21.0
 * Fixed compiler warnings for Fortran code.
 * Added glm_spec function.
+* Added add_noise_spec_snr function.
 
 # spant 2.20.0
 * Added vertical line colour and linetype options for stackplots.